+++ src/config.ts
@@ -15,6 +15,7 @@ export interface Config {
     autoPush: boolean;
     skipFetchGhPages: boolean;
     commentAlways: boolean;
+    summaryAlways: boolean;
     saveDataFile: boolean;
     commentOnAlert: boolean;
     alertThreshold: number;
@@ -34,6 +35,7 @@ export const VALID_TOOLS = [
     'googlecpp',
     'catch2',
     'julia',
+    'jmh',
     'benchmarkdotnet',
     'customBiggerIsBetter',
     'customSmallerIsBetter',
@@ -227,6 +229,7 @@ export async function configFromJobInput(): Promise<Config> {
     const autoPush = getBoolInput('auto-push');
     const skipFetchGhPages = getBoolInput('skip-fetch-gh-pages');
     const commentAlways = getBoolInput('comment-always');
+    const summaryAlways = getBoolInput('summary-always');
     const saveDataFile = getBoolInput('save-data-file');
     const commentOnAlert = getBoolInput('comment-on-alert');
     const alertThreshold = getPercentageInput('alert-threshold');
@@ -272,6 +275,7 @@ export async function configFromJobInput(): Promise<Config> {
         autoPush,
         skipFetchGhPages,
         commentAlways,
+        summaryAlways,
         saveDataFile,
         commentOnAlert,
         alertThreshold,